feature: dev 머지

+++ EduConnect/src/main/java/com/BugJava/EduConnect/assignment/service/AssignmentService.java
@@ -5,9 +5,13 @@ import com.BugJava.EduConnect.assignment.dto.AssignmentRequest;
 import com.BugJava.EduConnect.assignment.dto.AssignmentResponse;
 import com.BugJava.EduConnect.assignment.repository.AssignmentRepository;
 import com.BugJava.EduConnect.assignment.exception.PostNotFoundException;
+import com.BugJava.EduConnect.auth.entity.Users;
+import com.BugJava.EduConnect.auth.enums.Role;
+import com.BugJava.EduConnect.auth.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.security.access.AccessDeniedException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -18,15 +22,19 @@ import java.util.stream.Collectors;
 public class AssignmentService {
 
     private final AssignmentRepository assignmentRepository;
-
+    private final UserRepository userRepository;
     //과제 등록
     @Transactional
-    public AssignmentResponse createAssignment(AssignmentRequest request) {
-        //강사 권한 코드 추가 예정
-
+    public AssignmentResponse createAssignment(AssignmentRequest request, Long userId) {
+        Users user = userRepository.findById(userId)
+                .orElseThrow(() -> new PostNotFoundException("사용자를 찾을 수 없습니다."));
+        if (user.getRole() == Role.STUDENT) {
+            throw new AccessDeniedException("작성 권한이 없습니다.");
+        }
         Assignment assignment = Assignment.builder()
                 .title(request.getTitle())
                 .description(request.getDescription())
+                .user(user)
                 .build();
 
         Assignment saved = assignmentRepository.save(assignment);
@@ -36,7 +44,7 @@ public class AssignmentService {
     //전체 과제 목록 조회
     public List<AssignmentResponse> getAllAssignments() {
         return assignmentRepository.findAll().stream()
-                .map(AssignmentResponse::from)  // DTO의 from 메서드 사용
+                .map(AssignmentResponse::from)
                 .collect(Collectors.toList());
     }
 
@@ -50,9 +58,13 @@ public class AssignmentService {
 
     //과제 수정
     @Transactional
-    public AssignmentResponse updateAssignment(Long id, AssignmentRequest request) {
+    public AssignmentResponse updateAssignment(Long id, AssignmentRequest request, Long userId) {
         Assignment assignment = assignmentRepository.findById(id)
                 .orElseThrow(() -> new PostNotFoundException("과제를 찾을 수 없습니다." + id));
+
+        if (!assignment.getUser().getId().equals(userId)) {
+            throw new AccessDeniedException("본인이 작성한 과제만 수정할 수 있습니다.");
+        }
 
         if (request.getTitle() != null) {
             assignment.setTitle(request.getTitle());
@@ -66,10 +78,12 @@ public class AssignmentService {
 
     //과제 삭제
     @Transactional
-    public void deleteAssignment(Long id) {
+    public void deleteAssignment(Long id, Long userId) {
         Assignment assignment = assignmentRepository.findById(id)
                 .orElseThrow(() -> new PostNotFoundException("과제를 찾을 수 없습니다." + id));
-
+        if (!assignment.getUser().getId().equals(userId)) {
+            throw new AccessDeniedException("본인이 작성한 과제만 삭제할 수 있습니다.");
+        }
         assignmentRepository.delete(assignment);
     }
